--- backend/engine/plugins/github_repo_health/github_repo_health/src/__init__.py
@@ -1,2 +1,0 @@
-import rules
-from utilities import *
