+++ backend/engine/plugins/github_repo_health/lib/example_rules.json
