+++ backend/engine/plugins/github_repo_health/lib/src/github_repo_health/utilities/checker.py
@@ -1,4 +1,4 @@
-import rules
+from github_repo_health import rules
 
 
 class Checker:
