+++ backend/engine/plugins/github_repo_health/lib/tests/rules/test_branch_protection_commit_signing.py
@@ -1,8 +1,8 @@
 import unittest
 from unittest.mock import MagicMock
 
-from repohealth.rules import BranchProtectionCommitSigning
-from repohealth.utilities import Github
+from github_repo_health.rules import BranchProtectionCommitSigning
+from github_repo_health.utilities import Github
 
 OWNER = "owner"
 REPO = "repo"
