+++ backend/engine/plugins/github_repo_health/lib/tests/rules/test_branch_rule_bypass_actors.py
@@ -1,8 +1,8 @@
 import pytest
 from unittest.mock import MagicMock
 
-from repohealth.rules import BranchRulesetBypassActors
-from repohealth.utilities import Github
+from github_repo_health.rules import BranchRulesetBypassActors
+from github_repo_health.utilities import Github
 
 OWNER = "owner"
 REPO = "repo"
