+++ backend/engine/plugins/github_repo_health/lib/tests/rules/test_branch_rule_pull_requests.py
@@ -1,8 +1,8 @@
 import unittest
 from unittest.mock import MagicMock
 
-from repohealth.rules import BranchRulePullRequests
-from repohealth.utilities import Github
+from github_repo_health.rules import BranchRulePullRequests
+from github_repo_health.utilities import Github
 
 OWNER = "owner"
 REPO = "repo"
