+++ backend/engine/plugins/github_repo_health/lib/tests/rules/test_repo_code_scanning.py
@@ -1,8 +1,8 @@
 import unittest
 from unittest.mock import MagicMock
 
-from repohealth.rules import RepoCodeScanning
-from repohealth.utilities import Github
+from github_repo_health.rules import RepoCodeScanning
+from github_repo_health.utilities import Github
 
 OWNER = "owner"
 REPO = "repo"
